SageMath version 7.5.beta1

+++ src/sage/libs/pari/convert.pxd
@@ -8,6 +8,6 @@ cdef GEN gtoi(GEN g0) except NULL
 cdef GEN PyLong_AsGEN(x)
 cdef PyLong_FromGEN(GEN g)
 
-cdef gen new_t_POL_from_int_star(int *vals, int length, long varnum)
+cdef gen new_t_POL_from_int_star(int* vals, unsigned long length, long varnum)
 cdef gen new_gen_from_double(double)
 cdef gen new_t_COMPLEX_from_double(double re, double im)
